+++ game/cards/repository.go
@@ -347,4 +347,5 @@ var DM05 = map[string]match.CardConstructor{
 	"3e63d487-a25c-405d-8c04-df8c5e414fd6": dm05.SplitHeadHydroturtleQ,
 	"4ac7c097-cdc3-42e8-9203-1a481274f9d2": dm05.BladerushSkyterrorQ,
 	"5ad88bc8-5991-46d9-a98a-19c7569a4f05": dm05.RuthlessSkyterror,
+	"5f5a3a3f-3675-4e5d-8bd6-0c76cb8c4ce4": dm05.DeathCruzerTheAnnihilator,
 }